Assignment 7/Cloud_functions/gcs_to_bq/requirements.txt
google-cloud-storage
google-cloud-bigquery
google-cloud-pubsub
google-cloud-aiplatform
requests
db-dtypes
functions-framework